tests/wpt/web-platform-tests/css/css-transforms/parsing/perspective-origin-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: getComputedStyle().perspectiveOrigin</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#perspective-origin-property">
<meta name="assert" content="perspective-origin computed value is a pair of absolute lengths">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    width: 200px;
    height: 300px;
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("perspective-origin", "10%", "20px 150px");
test_computed_value("perspective-origin", "10% center", "20px 150px");
test_computed_value("perspective-origin", "20% 30px", "40px 30px");
test_computed_value("perspective-origin", "30px center", "30px 150px");
test_computed_value("perspective-origin", "40px top", "40px 0px");
test_computed_value("perspective-origin", "bottom 10% right 20%", "160px 270px");
test_computed_value("perspective-origin", "bottom right", "200px 300px");
test_computed_value("perspective-origin", "center", "100px 150px");
test_computed_value("perspective-origin", "center 50px", "100px 50px");
test_computed_value("perspective-origin", "center bottom", "100px 300px");
test_computed_value("perspective-origin", "center center", "100px 150px");
test_computed_value("perspective-origin", "center left", "0px 150px");
test_computed_value("perspective-origin", "left", "0px 150px");
test_computed_value("perspective-origin", "left 10px", "0px 10px");
test_computed_value("perspective-origin", "left bottom", "0px 300px");
test_computed_value("perspective-origin", "left center", "0px 150px");
test_computed_value("perspective-origin", "right 40%", "200px 120px");
test_computed_value("perspective-origin", "right 30% top -60px", "140px -60px");
test_computed_value("perspective-origin", "top", "100px 0px");
test_computed_value("perspective-origin", "right 20px bottom 30px", "180px 270px");

test_computed_value("perspective-origin", "right calc(10px - 0.5em) top calc(10px - 0.5em)", "210px -10px");
</script>
</body>
</html>
